Address review feedback on the coverage test additions
Deduplicate the observationConvention test into ObservationContract
via a conventionKueryClient factory method (mirroring
SqlIdGenerationDefaultContract's capturingKueryClient pattern), and
rename the new compiler checker tests to the third-person-singular
spec style used elsewhere in the suite.
